pwmの周期を受け取る変数の方を16bit整数に変更

+++ include/PwmOut.hpp
@@ -5,7 +5,7 @@
 struct PwmOut {
     uint8_t pinId;
     uint8_t channel;
-    uint8_t periodUs;
+    uint16_t periodUs;
     uint8_t dutyInt;
 
     void setup() {
